+++ src/router/index.ts
@@ -1,0 +1,4 @@
+// Optional re-export file in case we want to proxy in order to make the routes feel like a external lib
+
+export * from './components/RenderRoutes';
+export * from 'react-router-dom';
