--- src/routes/index.ts
@@ -1,3 +1,0 @@
-export * from './config';
-export * from './RenderRoutes';
-export * from 'react-router-dom';
