+++ src/shared/services/lazy-load/index.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 import React, { lazy, Suspense } from 'react';
 
 interface Opts {
