+++ src/theme/index.ts
@@ -1,0 +1,3 @@
+// this is // Optional re-export file in case we want to proxy in order to make the routes feel like a external lib
+
+export const theme = 'THEME';
